# public/doc/swagger.yaml
swagger: '2.0'
info:
  description: API Document
  version: 0.0.1
  title: Platform REST API
host: localhost:6041
basePath: /
tags:
  - name: document
    description: APIs for document(s)
schemes:
  - http
  - https
paths:
  /login:
    post:
      tags:
        - user
      summary: login
      description: Login authentication
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: Success
        '401':
          description: Access error
  /logout:
    post:
      tags:
        - user
      summary: logout
      description: Logout
      responses:
        '200':
          description: Success
        '401':
          description: Access error
  /doc/{id}:
    get:
      tags:
        - document
      summary: Get single document
      description: Get single document from id.
      security:
        - tokenAuth: []
      parameters:
        - name: id
          type: string
          in: path
          description: id of document
          required: true
      responses:
        '200':
          description: Success
        '401':
          description: Access error
    delete:
      tags:
        - document
      summary: Delete single document
      description: Delete single document from id.
      security:
        - tokenAuth: []
      parameters:
        - name: id
          type: string
          in: path
          description: id of document
          required: true
      responses:
        '200':
          description: Success
        '401':
          description: Access error
  /doc:
    post:
      tags:
        - document
      summary: Create document
      description: Create new document
      security:
        - tokenAuth: []
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/DocRequest'
      responses:
        '200':
          description: Success
        '401':
          description: Access error
  /docs:
    get:
      tags:
        - document
      summary: Get all documents
      description: Get all documents
      security:
        - tokenAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: Success
        '401':
          description: Access error
  /reset:
    post:
      tags:
        - document
      summary: Reset ledger
      description: Reset ledger
      security:
        - tokenAuth: []
      produces:
        - application/json
      responses:
        '200':
          description: Success
        '401':
          description: Access error
definitions:
  LoginRequest:
    type: object
    properties:
      username:
        type: string
        description: username
      password:
        type: string
        description: password
  DocRequest:
    type: object
    properties:
      name:
        type: string
        description: name
      age:
        type: integer
        description: age
securityDefinitions:
  tokenAuth:
    type: 'apiKey'
    name: 'Authorization'
    in: 'header'